fixed bugs

+++ src/edu/mu/PizzaOrder.java
@@ -68,9 +68,10 @@ public class PizzaOrder {
 		}
 		return true;
 	}
-	
-	//Method checks the pizzaOrderList for any uncooked pizza's
-	//if there are any uncooked pizza's then true is returned and otherwise false is returned
+
+	// Method checks the pizzaOrderList for any uncooked pizza's
+	// if there are any uncooked pizza's then true is returned and otherwise false
+	// is returned
 	public boolean isThereAnyUncookedPizza() {
 		for (AbstractPizza pizza : pizzaOrderList) {
 			if (pizza.getCookingStrategy() == null) {
@@ -79,40 +80,43 @@ public class PizzaOrder {
 		}
 		return false;
 	}
-	
-	//Method checkout calculates the total price in order list if there are no uncooked pizza's
-	//otherwise, an exception is thrown states that there are uncooked pizza's
+
+	// Method checkout calculates the total price in order list if there are no
+	// uncooked pizza's
+	// otherwise, an exception is thrown states that there are uncooked pizza's
 	public double checkout() throws Exception {
-		int total = 0;
-        if(isThereAnyUncookedPizza() == false) {
-        	for (AbstractPizza pizza : pizzaOrderList) {
-            	total += pizza.getTotalPrice();
-            	}
-        return total;
-        }
-        else {
-        	throw new Exception("Uncooked Pizzas");
-        }
+		double total = 0;
+		if (isThereAnyUncookedPizza() == false) {
+			for (AbstractPizza pizza : pizzaOrderList) {
+				total += pizza.getTotalPrice();
+			}
+			return total;
+		} else {
+			throw new Exception("Uncooked Pizzas");
+		}
 	}
-	
-	//Method takes an integer OrderID and a CookingStyleType for a pizza the method finds the pizza with that order id
-	//once the correct pizza is found, the method will setCookingStrategy and use the cookingStrategyType to invoke the correct cook() method
-	//Once the cooking is done, true is returned. else if the pizza order ID is not found then false is returned.s
+
+	// Method takes an integer OrderID and a CookingStyleType for a pizza the method
+	// finds the pizza with that order id
+	// once the correct pizza is found, the method will setCookingStrategy and use
+	// the cookingStrategyType to invoke the correct cook() method
+	// Once the cooking is done, true is returned. else if the pizza order ID is not
+	// found then false is returned.s
 	public boolean selectCookingStrategyByPizzaOrderID(int OrderID, CookingStyleType cookingStrategyType) {
 		for (AbstractPizza pizza : pizzaOrderList) {
 			if (pizza.getPizzaOrderID() == OrderID) {
-				switch(cookingStrategyType) {
-				  case MICROWAVE:
-				    MicrowaveCookingStrategy microwave = new MicrowaveCookingStrategy();
-				    microwave.cook(pizza);
-				    break;
-				  case CONVENTIONAL_OVEN:
-				    ConventionalOvenCookingStrategy conventionalOven = new ConventionalOvenCookingStrategy();
-				    conventionalOven.cook(pizza);
-				    break;
-				  default:
-				    BrickOvenCookingStrategy brickOven = new BrickOvenCookingStrategy();
-				    brickOven.cook(pizza);
+				switch (cookingStrategyType) {
+				case MICROWAVE:
+					MicrowaveCookingStrategy microwave = new MicrowaveCookingStrategy();
+					microwave.cook(pizza);
+					break;
+				case CONVENTIONAL_OVEN:
+					ConventionalOvenCookingStrategy conventionalOven = new ConventionalOvenCookingStrategy();
+					conventionalOven.cook(pizza);
+					break;
+				default:
+					BrickOvenCookingStrategy brickOven = new BrickOvenCookingStrategy();
+					brickOven.cook(pizza);
 				}
 				return true;
 			}
